Replace udev code with DBus code listening to events from iio-sensor-proxy
iio-sensor-proxy has underlying sensor code that closely resembles the removed accelerometer code. It is also a common program whose events many desktop environments await to do screen rotation. It also has other advantages in udev integration, such as a ruleset that controls when it is loaded and unloaded, and correctly handles devices with accelerometers oriented differently from their screens and other edge cases. Finally, it is likely to be more optimized, and receives much more testing and scrutiny than this script does. There are potential downsides in that iio-sensor-proxy seems to explicitly require systemd. Frankly, I am unsure whether this script once worked without systemd, because I have never used udev without systemd, but, anyone for whom this is a concern is advised not to update the script.